requirements.txt
requests==2.33.1
pandas==3.0.2
python-dotenv==1.2.2
